+++ notification-service/src/main/java/com/twitter/notification/service/repository/NotificationRepository.java
@@ -7,10 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface NotificationRepository extends JpaRepository<Notification, Integer> {
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
 
     List<Notification>
-    findByReceiverUserIdOrderByCreatedAtDesc(
-            Integer receiverUserId
-    );
+    findByReceiverUserIdOrderByCreatedAtDesc( Long receiverUserId );
 }
